+++ lib/titan/thread.rb
@@ -119,7 +119,7 @@ module Titan
       # Saves threads to pid files inside the TITAN_DIRECTORY
       #
       def save_threads
-        pid_files.each { |pid_file| File.delete (File.expand_path(pid_file, TITAN_DIRECTORY)) }
+        pid_files.each { |pid_file| File.delete(File.expand_path(pid_file, TITAN_DIRECTORY)) }
         @@threads.each_value{ |thread| thread.save }
       end